refactor: extract property domain types to types/
