+++ frontend/app/listings/page.tsx
@@ -4,7 +4,8 @@ import { useEffect, useState } from "react";
 import { usePathname } from "next/navigation";
 
 import { ButtonPrimary } from "@components/ui/button-primary";
-import { detailToModel, type PropertyModel } from "@components/property/listing-model";
+import { detailToModel } from "@components/property/listing-model";
+import type { PropertyModel } from "@typings/property";
 import { PropertyCard, PropertyCardSkeleton, PROPERTY_GRID } from "@components/property/property-card";
 import { Heading2 } from "@components/ui/heading2";
 import { brand } from "@config/brand";
