+++ frontend/components/property/section-grid-feature-property.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from "react";
 
 import { ButtonSecondary } from "@components/ui/button-secondary";
 import { Heading2 } from "@components/ui/heading2";
-import type { PropertyModel } from "@components/property/listing-model";
+import type { PropertyModel } from "@typings/property";
 import { PropertyCard, PROPERTY_GRID } from "@components/property/property-card";
 import { cn } from "@utils/common";
