+++ frontend/components/saved/view.tsx
@@ -5,7 +5,8 @@ import { useEffect, useState } from "react";
 import { useSavedListings } from "@components/saved/provider";
 import { NoticeCard } from "@components/ui/notice-card";
 import { ButtonPrimary } from "@components/ui/button-primary";
-import { detailToModel, type PropertyModel } from "@components/property/listing-model";
+import { detailToModel } from "@components/property/listing-model";
+import type { PropertyModel } from "@typings/property";
 import {
   PropertyCard,
   PropertyCardSkeleton,
